@@ -4,10 +4,12 @@
     "type": "library",
     "license": "MIT",
     "require": {
-        "illuminate/support": "^5.7|^5.8|^5.9|^6.0|^7.0",
-        "php": "^7.2",
-        "illuminate/database": "^5.7|^6.0|^7.0",
-        "illuminate/http": "^5.7|^6.0|^7.0"
+        "illuminate/support": "^5.7|^5.8|^5.9|^6.0|^7.0|^8.0",
+        "php": "^7.4",
+        "illuminate/database": "^5.7|^6.0|^7.0|^8.0",
+        "illuminate/http": "^5.7|^6.0|^7.0|^8.0",
+        "illuminate/pipeline": "^5.7|^6.0|^7.0|^8.0",
+        "illuminate/console": "^5.7|^6.0|^7.0|^8.0"
     },
     "require-dev": {
         "phpunit/phpunit": "^8.1"
